itau-extrato-db/requirements.txt
pypdf==4.3.1
python-dateutil==2.9.0.post0
Unidecode==1.3.8
fastapi==0.115.0
uvicorn==0.30.6
Jinja2==3.1.4
